Simplified the structure
