+++ server/controllers/user.controller.ts
@@ -1,15 +1,14 @@
-const userService = require('../services/User.service');
+import { userInfo } from "os";
+
 const {JWT} = require('../token/jwt');
+const bcrypt = require('bcrypt');
+const database = require('../db');
+const router = express.Router();
 
 class UserController {
-     userService: any;
-
-     constructor (userService: any){
-         this.userService = userService;
-     }
 
      async login(req, res){
-        const user = await this.userService.getUserByEmail(req.body.email);
+        const user = await this.getUserByEmail(req.body.email);
 
         if (!user){
             return res.status(400).json({
@@ -17,7 +16,7 @@ class UserController {
             });
         }
 
-        const validUser = await this.userService.verifyUser(req.body.password, user.password);
+        const validUser = await this.verifyUser(req.body.password, user.password);
         if (!user){
             return res.status(400).json({
                 message: 'Invalid password!',
@@ -44,6 +43,32 @@ class UserController {
             id: user.id,
      });
     }
+    
+    async getUserByEmail(email: string){
+        const SQL = `SELECT * FROM user WHERE user.email = '${email}'`; 
+        const client = database.connect();
+        try {
+            const res = await database.client.query(SQL);
+            return res;
+        }
+        catch(err){
+            console.error(err);
+        }
+        finally {
+            // Make sure to release the client before any error handling,
+            // just in case the error handling itself throws an error.
+            client.release()
+          }
+    }
+
+    async verifyUser (candidatePassword:string, encryptedPass:any) {
+        return bcrypt.compare(candidatePassword, encryptedPass);
+    }
 }
 
-module.exports = new UserController(userService);
+router.get('/', (req, res) => {
+    const controller = new UserController();
+    return controller.login(req, res);
+});
+
+module.exports = router;
