+++ server/db/Database.ts
@@ -5,8 +5,6 @@ const { Pool }  = require('pg');
 class Database {
     connConfig: { user: string; host: string; database: string; password: string; port: number; };
     pool: any;
-    client: Promise<any>;
-
     constructor() {
         this.connConfig = {
             user: process.env.POSTGRES_USER,
@@ -17,20 +15,10 @@ class Database {
           };
           
         this.pool = new Pool(this.connConfig);
-        this.client = this.connect();
     }
 
-    async connect () {
-        try {
-            var client = await this.pool.connect();
-          
-            console.log('[DB]: Successfully connected.');
-            return client;
-
-          } catch (err: any) {
-            console.log(err.message);
-            return null;
-          }
+    async getPool () {
+        return this.pool;
     }
 }
