+++ server/index.ts
@@ -1,5 +1,5 @@
 const express = require('express');
-const login = require(''); // there it goes the login page i guess?
+const login = require('./controllers/user.controller');
 const db = require('../db');
 
 const app = express();
@@ -9,23 +9,26 @@ app.use(express.json());
 
 app.get('/test', (req, res) => res.json({ message: 'this is a message!' }));
 
+app.use('/login', login); 
+
 app.post('/save-trip',async(req, res) => {
 
   const trip = req.body.trip;
   const activities = req.body.activities; 
+  const client = await db.getPool().connect();
   try {
     const values = Object.values(trip)
-    let res = await db.client.query(`INSERT INTO trip(${Object.keys(trip).join(', ')}) VALUES($1, $2, $3, $4, $5) RETURNING id`, values)
+    let res = await client.query(`INSERT INTO trip(${Object.keys(trip).join(', ')}) VALUES($1, $2, $3, $4, $5) RETURNING id`, values)
     const id = res.rows[0].id;
     const activities_values = activities.map((a:any) => ({...a, trip_id:id}));
-    res = await db.client.query(`INSERT INTO trip_activity(${Object.keys(activities_values[0]).join(', ')}) VALUES ${activities_values.map((a:any) => `(${Object.values(a)
+    res = await client.query(`INSERT INTO trip_activity(${Object.keys(activities_values[0]).join(', ')}) VALUES ${activities_values.map((a:any) => `(${Object.values(a)
       .map(v => typeof v === "string" ? `'${v}'` : v)
     .join(', ')})`)}`)
     console.log(res)
   } finally {
     // Make sure to release the client before any error handling,
     // just in case the error handling itself throws an error.
-    db.client.release()
+    client.release()
   }
   
   res.json({ message: 'trip sucessfully added!' })
