--- server/services/index.ts
@@ -1,8 +1,0 @@
-class Service {
-
-    constructor (name: string) {
-        // something ???
-    }
-}
-
-module.exports = Service;
